Update README.md
This is a form to populate a case by case response to common task and ticket request. 

Choose generic subject enter case specific properties into a response paragraph that can be pasted and edited in an email ticketing system.

custom user intros and closing staments for real feel or un-robotic perceptions from client.